requirements.txt
torch==2.0.1
transformers==4.31.0

torchvision==0.15.2
Pillow==10.0.0 
pydicom==2.4.3 

uvicorn==0.23.2
fastapi==0.101.1 

python-multipart==0.0.6 